🛠 refactor: entities 폴더 구조 변경
모듈 종속 ➡ src 폴더 내 최상위로 계층 변경
