+++ src/auth/outbound-adapter/create-user.repository.ts
@@ -1,7 +1,7 @@
+import { Profile } from './../../entities/profile.entity';
+import { User } from './../../entities/user.entity';
 import { Injectable, InternalServerErrorException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { User } from './../../user/entities/user.entity';
-import { Profile } from '../../profile/entities/profile.entity';
 import { Repository } from 'typeorm';
 import {
   CreateUserOutboundPort,
