+++ src/auth/outbound-adapter/get-user-by-email.repository.ts
@@ -1,6 +1,6 @@
+import { User } from './../../entities/user.entity';
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { User } from './../../user/entities/user.entity';
 import { Repository } from 'typeorm';
 import {
   GetUserByEmailOutboundPort,
