+++ src/user/dtos/update-user.dto.ts
@@ -1,0 +1,3 @@
+import { PartialType } from '@nestjs/swagger';
+
+// export class UpdateUserDto extends PartialType() {}
