+++ src/user/types/user.repository.ts
@@ -1,5 +1,5 @@
+import { User } from './../../entities/user.entity';
 import { CreateUserDto } from '../dtos/create-user.dto';
-import { User } from '../entities/user.entity';
 
 export class IUserRepository {
   create: (signupDto: CreateUserDto) => Promise<User>;
